requirements.txt
torch
torchvision
opencv-python
numpy
streamlit
plotly
Pillow
matplotlib
seaborn 
scikit-learn